Fix overflow in get_token_alloc leading to OOM
A signed integer overflow occurs in get_token_alloc() when the
VP9 encoder is initialized with large frame dimensions.
This causes incorrect buffer size calculation and OOM.

This commit fixes the overflow by using 64-bit arithmetic
in get_token_alloc() and updating callers to avoid truncation.

Added a unit test LargeDimensionsTokenAllocOverflow to verify
that large dimensions are handled without crashing or OOMing.

Bug: 505286908

+++ vp9/encoder/vp9_encoder.c
@@ -1393,7 +1393,7 @@ static void alloc_compressor_data(VP9_COMP *cpi) {
   vpx_free(cpi->tile_tok[0][0]);
 
   {
-    unsigned int tokens = get_token_alloc(cm->mb_rows, cm->mb_cols);
+    size_t tokens = get_token_alloc(cm->mb_rows, cm->mb_cols);
     CHECK_MEM_ERROR(&cm->error, cpi->tile_tok[0][0],
                     vpx_calloc(tokens, sizeof(*cpi->tile_tok[0][0])));
   }
